Use helium-js solana package.
